no-op implementation of spawning on non-windows platforms
